Stock_Market/requirements.txt
streamlit
plotly
yfinance
newsapi-python
textblob
scikit-learn
pandas
streamlit-autorefresh
